Create shared base class for different asset VC's
