+++ ImageViewer/AssetViewController/VideoViewerController/VideoViewController.xib
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.XIB" version="3.0" toolsVersion="14109" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" colorMatched="YES">
+    <device id="retina4_7" orientation="portrait">
+        <adaptation id="fullscreen"/>
+    </device>
+    <dependencies>
+        <deployment identifier="iOS"/>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="14088"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <objects>
+        <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="VideoViewController" customModule="SimpleImageViewer" customModuleProvider="target">
+            <connections>
+                <outlet property="assetView" destination="7UU-24-FEq" id="mjT-6Q-MZe"/>
+                <outlet property="scrollView" destination="KG2-ZX-S5t" id="aEr-lm-oGA"/>
+                <outlet property="view" destination="iN0-l3-epB" id="FTF-w5-1qm"/>
+            </connections>
+        </placeholder>
+        <placeholder placeholderIdentifier="IBFirstResponder" id="-2" customClass="UIResponder"/>
+        <view contentMode="scaleToFill" id="iN0-l3-epB">
+            <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+            <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+            <subviews>
+                <scrollView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" contentInsetAdjustmentBehavior="never" maximumZoomScale="6" translatesAutoresizingMaskIntoConstraints="NO" id="KG2-ZX-S5t">
+                    <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                    <subviews>
+                        <view contentMode="scaleToFill" translatesAutoresizingMaskIntoConstraints="NO" id="7UU-24-FEq">
+                            <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+                        </view>
+                    </subviews>
+                    <constraints>
+                        <constraint firstItem="7UU-24-FEq" firstAttribute="leading" secondItem="KG2-ZX-S5t" secondAttribute="leading" id="87i-jg-dpS"/>
+                        <constraint firstAttribute="trailing" secondItem="7UU-24-FEq" secondAttribute="trailing" id="9hT-7U-vsG"/>
+                        <constraint firstAttribute="bottom" secondItem="7UU-24-FEq" secondAttribute="bottom" id="J1q-WM-VXV"/>
+                        <constraint firstItem="7UU-24-FEq" firstAttribute="top" secondItem="KG2-ZX-S5t" secondAttribute="top" id="fld-GD-xgj"/>
+                    </constraints>
+                    <connections>
+                        <outlet property="delegate" destination="-1" id="606-e2-Lyo"/>
+                    </connections>
+                </scrollView>
+            </subviews>
+            <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="calibratedWhite"/>
+            <constraints>
+                <constraint firstItem="7UU-24-FEq" firstAttribute="height" secondItem="iN0-l3-epB" secondAttribute="height" id="07L-nQ-ndy"/>
+                <constraint firstAttribute="trailing" secondItem="KG2-ZX-S5t" secondAttribute="trailing" id="3PU-ep-Ehf"/>
+                <constraint firstItem="KG2-ZX-S5t" firstAttribute="leading" secondItem="iN0-l3-epB" secondAttribute="leading" id="9aH-SN-GDp"/>
+                <constraint firstItem="7UU-24-FEq" firstAttribute="width" secondItem="iN0-l3-epB" secondAttribute="width" id="bDV-Jz-1sp"/>
+                <constraint firstAttribute="bottom" secondItem="KG2-ZX-S5t" secondAttribute="bottom" id="t0V-5N-tGM"/>
+                <constraint firstItem="KG2-ZX-S5t" firstAttribute="top" secondItem="iN0-l3-epB" secondAttribute="top" id="wbc-D6-6Mx"/>
+            </constraints>
+            <point key="canvasLocation" x="33.5" y="72.5"/>
+        </view>
+    </objects>
+</document>
